fix(DF-54): 리뷰 관련 페이지의 백버튼 로직 수정 - 일부 페이지에서 스타일 코드 분리
