+++ src/pages/main/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState, ChangeEvent, useEffect } from "react";
-import styled from "styled-components";
 import { AxiosError } from "axios";
+import S from "./main.styles";
 import { MainMap } from "../../components/map/MainMap";
 import BottomSheet from "../../components/bottomsheet/BottomSheet";
 import BottomSheetHeader from "./components/BottomSheetHeader";
@@ -21,126 +21,6 @@ import { useNavigate } from "react-router-dom";
 import { useLocationStore } from "../../store/useLocationStore";
 import { MdOutlineInbox } from "react-icons/md";
 import SearchBar from "./components/SearchInput";
-
-const MainContainer = styled.div`
-  position: relative;
-  height: 100vh;
-  width: 100%;
-`;
-
-const SearchBarContainer = styled.div`
-  position: fixed;
-  left: 50%;
-  top: 20px;
-  transform: translateX(-50%);
-  width: 90%;
-  max-width: 400px;
-  z-index: 10;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  /* 태블릿 환경 */
-  @media screen and (min-width: 700px) {
-    max-width: 100%;
-    top: 25px;
-    gap: 12px;
-  }
-
-  /* 큰 태블릿 및 노트북 */
-  @media screen and (min-width: 1024px) {
-    max-width: 600px;
-    top: 30px;
-  }
-`;
-
-const BottomSheetContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  width: 100%;
-`;
-
-const FixedHeader = styled.div`
-  position: sticky;
-  top: 0;
-  background: white;
-  z-index: 1;
-  flex-shrink: 0;
-`;
-
-const PathCardList = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-  padding: 10px 0 70px 0;
-  overflow-y: auto;
-  flex: 1;
-  -webkit-overflow-scrolling: touch;
-`;
-
-const ErrorMessage = styled.div`
-  text-align: center;
-  padding: 2rem;
-  color: ${theme.Black};
-  font-size: 1rem;
-`;
-
-const LoadingSpinner = styled.div`
-  text-align: center;
-  padding: 1rem;
-`;
-
-const EmptyStateContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 20px 0;
-  width: 100%;
-  /* 태블릿 환경 */
-  @media screen and (min-width: 700px) {
-    margin: 30px 0;
-  }
-
-  /* 큰 태블릿 및 노트북 */
-  @media screen and (min-width: 1024px) {
-    margin: 40px 0;
-  }
-`;
-
-const IconWrapper = styled.div`
-  background-color: ${theme.Gray100};
-  border-radius: 50%;
-  width: 80px;
-  height: 80px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 16px;
-`;
-
-const NoWalkwaysMessage = styled.p`
-  font-size: 16px;
-  color: ${theme.Gray700};
-  text-align: center;
-  line-height: 1.5;
-  margin: 0;
-  font-weight: 500;
-`;
-
-const ViewAllButton = styled.button`
-  background-color: ${theme.Green300};
-  color: white;
-  border: none;
-  border-radius: 8px;
-  padding: 10px 25px;
-  font-size: 14px;
-  font-weight: 500;
-  margin-top: 30px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-`;
 
 interface Location {
   lat: number;
@@ -536,8 +416,8 @@ function Main() {
 
   return (
     <>
-      <MainContainer>
-        <SearchBarContainer>
+      <S.MainContainer>
+        <S.SearchBarContainer>
           <GuideButton onClick={() => navigate("/guide")} />
           <SearchBar
             value={searchValue}
@@ -549,8 +429,8 @@ function Main() {
             onSelect={handleResultSelect}
             onOutsideClick={() => setSearchResults([])}
           />
-        </SearchBarContainer>
-
+        </S.SearchBarContainer>
+  
         <MainMap
           center={
             selectedLocation
@@ -569,7 +449,7 @@ function Main() {
           onLocationButtonClick={handleLocationButtonClick}
           onSearchCurrentLocation={handleSearchCurrentLocation}
         />
-
+  
         <BottomSheet
           isOpen={isOpen}
           onClose={toggleBottomSheet}
@@ -578,18 +458,18 @@ function Main() {
           showPreview={true}
           closeOnOutsideClick={true}
         >
-          <BottomSheetContainer>
-            <FixedHeader>
+          <S.BottomSheetContainer>
+            <S.FixedHeader>
               <BottomSheetHeader
                 sortValue={sortOption}
                 mapValue={mapOption}
                 onSortChange={handleSortChange}
                 onMapChange={handleMapChange}
               />
-            </FixedHeader>
-            <PathCardList onScroll={handleScroll}>
+            </S.FixedHeader>
+            <S.PathCardList onScroll={handleScroll}>
               {error ? (
-                <ErrorMessage>{error}</ErrorMessage>
+                <S.ErrorMessage>{error}</S.ErrorMessage>
               ) : walkways.length > 0 ? (
                 walkways.map((walkway) => (
                   <PathCard
@@ -615,36 +495,36 @@ function Main() {
               ) : (
                 !loading && (
                   <>
-                    <EmptyStateContainer>
-                      <IconWrapper>
+                    <S.EmptyStateContainer>
+                      <S.IconWrapper>
                         <MdOutlineInbox size={40} color={theme.Gray500} />
-                      </IconWrapper>
-                      <NoWalkwaysMessage>
+                      </S.IconWrapper>
+                      <S.NoWalkwaysMessage>
                         {mapOption === "current"
                           ? "전방 500m 부근에 등록된 산책로가 없습니다."
                           : "등록된 산책로가 없습니다."}
-                      </NoWalkwaysMessage>
-
+                      </S.NoWalkwaysMessage>
+  
                       {/* 현재 위치 모드에서만 버튼 표시 */}
                       {mapOption === "current" && (
-                        <ViewAllButton
+                        <S.ViewAllButton
                           onClick={() => {
                             setMapOption("all");
                             fetchAllWalkways(sortOption, true);
                           }}
                         >
                           동산의 모든 산책로 둘러보기
-                        </ViewAllButton>
+                        </S.ViewAllButton>
                       )}
-                    </EmptyStateContainer>
+                    </S.EmptyStateContainer>
                   </>
                 )
               )}
-              {loading && <LoadingSpinner />}
-            </PathCardList>
-          </BottomSheetContainer>
+              {loading && <S.LoadingSpinner />}
+            </S.PathCardList>
+          </S.BottomSheetContainer>
         </BottomSheet>
-      </MainContainer>
+      </S.MainContainer>
       <BottomNavigation />
     </>
   );
